+++ src/containers/User/Detail/Account/components/AdminSection.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { isUserAdmin } from "../../helpers";
+import InfoSection from "../../InfoSection.jsx";
+
+export default ({ isadmin }) =>
+  <InfoSection
+    className="no-margin-top"
+    label="Admin:"
+    info={isUserAdmin(isadmin) ? "Yes" : "No"}
+  />;
