+++ src/containers/User/Detail/Account/components/EmailSection.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import { isUserEmailVerified } from "../../helpers";
+import InfoSection from "../../InfoSection.jsx";
+
+export default ({ token }) =>
+  <InfoSection
+    label="Verified email:"
+    info={isUserEmailVerified(token) ? "Yes" : "No"}
+  />;
