+++ src/containers/User/Detail/Identity/components/PastKeysSection.jsx
@@ -1,0 +1,46 @@
+import React from "react";
+import { classNames, Text, Button, P, Modal } from "pi-ui";
+import styles from "../Identity.module.css";
+import useBooleanState from "src/hooks/utils/useBooleanState";
+import IdentityList from "./IdentityList";
+
+export default ({ pastIdentities }) => {
+  const [
+    showShowAllModal,
+    openShowAllModal,
+    closeShowAllModal
+  ] = useBooleanState(false);
+  return (
+    <>
+      <Text
+        color="grayDark"
+        weight="semibold"
+        className={classNames(
+          styles.fieldHeading,
+          "margin-bottom-s",
+          "margin-top-l"
+        )}
+      >
+        Past public keys
+      </Text>
+      <P>
+        {pastIdentities.length !== 0
+          ? "List of inactive public keys your account has had in the past."
+          : "This account only had one active public key until now."}
+      </P>
+      {pastIdentities.length !== 0 && (
+        <Button size="sm" kind="primary" onClick={openShowAllModal}>
+          Show all
+        </Button>
+      )}
+      <Modal
+        show={showShowAllModal}
+        onClose={closeShowAllModal}
+        title="Past public keys"
+      >
+        <IdentityList full identities={pastIdentities} />
+      </Modal>
+    </>
+  );
+}
+;
